requirements/base.txt
requests
numpy
pandas
seaborn 
matplotlib
sp-ask-school

django==4.2.16
django-crispy-forms
crispy-bootstrap4
calplot

python-dateutil 
python-dotenv
BeautifulSoup4
colorama==0.4.4

flower==1.2.0
#psycopg2-binary==2.9.2 # new
redis==4.0.2

tinydb

#deploy
fabric==2.7.0
fabric2==2.7.0

#libraryH3lp API Client
#https://gitlab.com/libraryh3lp/libraryh3lp-sdk-python/-/tree/master
lh3api

#stats
#Create MS Excel file 
XlsxWriter
openpyxl

#dev
django-extensions   

django-model-utils


#celery
#celery
#django-celery-beat
#django-celery-results
#celery_once

pytz
gunicorn
amqp

isort==4.3.21

argh==0.26.2
PyYAML==6.0
watchdog==2.1.6

pyclean

setuptools>=70.0.0


nltk
scikit-learn


